requirements.txt
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==0.19.0
pytz==2024.1
requests==2.32.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.0
WTForms==3.0.1
